move DocumentActionBadge to documents folder
