+++ packages/frontend/src/components/documents/index.js
@@ -2,10 +2,12 @@ import DocumentTotalCard from './DocumentTotalCard'
 import DocumentsList from './DocumentsList'
 import DocumentsListItem from './DocumentsListItem'
 import DocumentsRevisionsList from './revisions/DocumentsRevisionsList'
+import DocumentActionBadge from './DocumentActionBadge'
 
 export {
   DocumentTotalCard,
   DocumentsList,
   DocumentsListItem,
-  DocumentsRevisionsList
+  DocumentsRevisionsList,
+  DocumentActionBadge
 }
